+++ AzeriteUI/Components/Misc/ArcheologyBar.lua
@@ -190,7 +190,6 @@ end
 
 ArcheologyBar.OnInitialize = function(self)
 	self.db = ns.db:RegisterNamespace("ArcheologyBar", defaults)
-	--self.db:SetProfile("Default")
 
 	self:SetEnabledState(self.db.profile.enabled)