+++ AzeriteUI/Components/Misc/ObjectivesTracker.lua
@@ -663,7 +663,6 @@ end
 
 Tracker.OnInitialize = function(self)
 	self.db = ns.db:RegisterNamespace("ObjectivesTracker", defaults)
-	--self.db:SetProfile("Default")
 
 	self:SetEnabledState(self.db.profile.enabled)
 	self:HookTracker()