+++ AzeriteUI/Components/Misc/RaidWarnings.lua
@@ -209,7 +209,6 @@ end
 
 RaidWarnings.OnInitialize = function(self)
 	self.db = ns.db:RegisterNamespace("RaidWarnings", defaults)
-	--self.db:SetProfile("Default")
 
 	self:SetEnabledState(self.db.profile.enabled)