+++ AzeriteUI/Components/Misc/VehicleSeat.lua
@@ -186,7 +186,6 @@ end
 
 VehicleSeat.OnInitialize = function(self)
 	self.db = ns.db:RegisterNamespace("VehicleSeat", defaults)
-	--self.db:SetProfile("Default")
 
 	self:SetEnabledState(self.db.profile.enabled)