+++ AzeriteUI/Components/UnitFrames/Units/Boss.lua
@@ -629,7 +629,6 @@ end
 
 BossFrameMod.OnInitialize = function(self)
 	self.db = ns.db:RegisterNamespace("BossFrames", defaults)
-	--self.db:SetProfile("Default")
 
 	self:SetEnabledState(self.db.profile.enabled)