+++ AzeriteUI/Components/UnitFrames/Units/Focus.lua
@@ -484,7 +484,6 @@ end
 
 FocusFrameMod.OnInitialize = function(self)
 	self.db = ns.db:RegisterNamespace("FocusFrame", defaults)
-	--self.db:SetProfile("Default")
 
 	self:SetEnabledState(self.db.profile.enabled)